# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SaaSKit Multi-Application API
  version: 0.4.0
  description: Self-hosted multi-application SaaS API. Users and billing data are application-scoped; Alipay and WeChat merchant settings are deployment-global.
servers: [{url: http://localhost:8080/api}]
...
paths:
...
  /user-auth/register:
    post:
      tags: [UserAuth]
      security: [{appKey: []}]
      summary: Register a first-party SaaS user
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/UserCreate'}}}}
      responses: {'201': {description: User and user access token}, '403': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}
  /user-auth/login:
    post:
      tags: [UserAuth]
      security: [{appKey: []}]
      summary: SaaS user login
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [email, password], properties: {email: {type: string, format: email}, password: {type: string}}}}}}
      responses: {'200': {description: User and user access token}, '401': {$ref: '#/components/responses/Error'}, '403': {$ref: '#/components/responses/Error'}}
  /account/profile:
    get: {tags: [UserAuth], security: [{userBearer: []}], summary: Get current user profile, responses: {'200': {description: User profile}}}
    put:
      tags: [UserAuth]
      security: [{userBearer: []}]
      summary: Update current user profile
      requestBody: {required: true, content: {application/json: {schema: {type: object, properties: {name: {type: string}, phone: {type: string}, avatar_url: {type: string}}}}}}
      responses: {'200': {description: Updated user}}
  /account/password:
    put:
      tags: [UserAuth]
      security: [{userBearer: []}]
      summary: Change current user password
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [current_password, new_password], properties: {current_password: {type: string}, new_password: {type: string, minLength: 8}}}}}}
      responses: {'200': {description: Password changed}, '401': {$ref: '#/components/responses/Error'}}
...
  /admin/users:
    get:
      tags: [Users]
      security: [{adminBearer: [], appId: []}]
      summary: Search and paginate owned users
      parameters: [{name: q, in: query, schema: {type: string}}, {name: status, in: query, schema: {type: string, enum: [active, disabled]}}, {name: page, in: query, schema: {type: integer, minimum: 1}}, {name: page_size, in: query, schema: {type: integer, minimum: 1, maximum: 100}}]
      responses: {'200': {description: Paginated users}}
    post:
      tags: [Users]
      security: [{adminBearer: [], appId: []}]
      summary: Create user from administration
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/UserCreate'}}}}
      responses: {'201': {description: Created user}, '409': {$ref: '#/components/responses/Error'}}
  /admin/users/{userId}:
    parameters: [{$ref: '#/components/parameters/UserId'}]
    get: {tags: [Users], security: [{adminBearer: [], appId: []}], summary: Get user, responses: {'200': {description: User}, '404': {$ref: '#/components/responses/Error'}}}
    put:
      tags: [Users]
      security: [{adminBearer: [], appId: []}]
      summary: Update user identity and verification fields
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/UserUpdate'}}}}
      responses: {'200': {description: Updated user}, '409': {$ref: '#/components/responses/Error'}}
    delete: {tags: [Users], security: [{adminBearer: [], appId: []}], summary: Soft-delete a user without orders, responses: {'200': {description: Deleted}, '409': {$ref: '#/components/responses/Error'}}}
  /admin/users/{userId}/status:
    parameters: [{$ref: '#/components/parameters/UserId'}]
    put:
      tags: [Users]
      security: [{adminBearer: [], appId: []}]
      summary: Enable or disable user login
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [status], properties: {status: {type: string, enum: [active, disabled]}}}}}}
      responses: {'200': {description: Updated user}}
  /admin/users/{userId}/password:
    parameters: [{$ref: '#/components/parameters/UserId'}]
    put:
      tags: [Users]
      security: [{adminBearer: [], appId: []}]
      summary: Reset user password
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [password], properties: {password: {type: string, minLength: 8}}}}}}
      responses: {'200': {description: Password reset}}
...
components:
  securitySchemes:
    adminBearer: {type: http, scheme: bearer, bearerFormat: JWT}
    userBearer: {type: http, scheme: bearer, bearerFormat: JWT}
    apiKey: {type: apiKey, in: header, name: X-API-Key}
    apiSecret: {type: apiKey, in: header, name: X-API-Secret}
    appId: {type: apiKey, in: header, name: X-App-ID, description: Required by application-scoped admin endpoints}
    appKey: {type: apiKey, in: header, name: X-App-Key, description: Required by public user registration and login}
  parameters:
    UserId: {name: userId, in: path, required: true, schema: {type: string, format: uuid}}
...
  responses:
    Error: {description: Error envelope, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}}
  schemas:
    ErrorEnvelope: {type: object, required: [code, message], properties: {code: {type: integer}, message: {type: string}}}
...
    UserCreate: {type: object, required: [email, password], properties: {email: {type: string, format: email}, password: {type: string, minLength: 8}, name: {type: string}, phone: {type: string}, status: {type: string, enum: [active, disabled]}}}
    UserUpdate: {type: object, properties: {email: {type: string, format: email}, name: {type: string}, phone: {type: string}, avatar_url: {type: string}, email_verified: {type: boolean}, phone_verified: {type: boolean}}}
...